requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
pymongo==4.13.1
aiofiles==25.1.0